# perf: dimension-based kernel for L2 and Cosine