--- a/rust/lance-linalg/src/distance/l2.rs
+++ b/rust/lance-linalg/src/distance/l2.rs
@@ -18,7 +18,7 @@
 use std::iter::Sum;
 use std::sync::Arc;
 
-use crate::distance::l2::f32::{l2_f32x16, l2_f32x8};
+use crate::distance::l2::f32::l2_once;
 use arrow_array::{cast::AsArray, types::Float32Type, Array, FixedSizeListArray, Float32Array};
 use half::{bf16, f16};
 use num_traits::real::Real;
@@ -136,21 +136,14 @@ pub fn l2_distance(from: &[f32], to: &[f32]) -> f32 {
 mod f32 {
     use super::*;
 
-    #[inline]
-    pub(crate) fn l2_f32x8(x: &[f32], y: &[f32]) -> f32 {
-        debug_assert_eq!(x.len(), 16);
-        debug_assert_eq!(y.len(), 8);
-        let x = unsafe { f32x16::load_unaligned(x.as_ptr()) };
-        let y = unsafe { f32x16::load_unaligned(y.as_ptr()) };
-        let s = x - y;
-        (s * s).reduce_sum()
-    }
+    use num_traits::Float;
 
     #[inline]
-    pub(crate) fn l2_f32x16(x: &[f32], y: &[f32]) -> f32 {
-        debug_assert_eq!(y.len(), 16);
-        let x = unsafe { f32x16::load_unaligned(x.as_ptr()) };
-        let y = unsafe { f32x16::load_unaligned(y.as_ptr()) };
+    pub(crate) fn l2_once<S: SIMD<F, N>, F: Float, const N: usize>(x: &[F], y: &[F]) -> F {
+        debug_assert_eq!(x.len(), S::LANES);
+        debug_assert_eq!(y.len(), S::LANES);
+        let x = unsafe { S::load_unaligned(x.as_ptr()) };
+        let y = unsafe { S::load_unaligned(y.as_ptr()) };
         let s = x - y;
         (s * s).reduce_sum()
     }
@@ -177,8 +170,14 @@ pub fn l2_distance_batch<'a>(
 
     // Dispatch based on the dimension.
     match dimension {
-        8 => Box::new(to.chunks_exact(dimension).map(move |v| l2_f32x8(from, v))),
-        16 => Box::new(to.chunks_exact(dimension).map(move |v| l2_f32x16(from, v))),
+        8 => Box::new(
+            to.chunks_exact(dimension)
+                .map(move |v| l2_once::<f32x8, f32, 8>(from, v)),
+        ),
+        16 => Box::new(
+            to.chunks_exact(dimension)
+                .map(move |v| l2_once::<f32x16, f32, 16>(from, v)),
+        ),
         _ => Box::new(to.chunks_exact(dimension).map(|v| from.l2(v))),
     }
 }