… rust/lance-linalg/src/distance/x86_64.rs
… rust/lance-linalg/src/kmeans.rs
… rust/lance-linalg/src/simd.rs