Provide faster routine for large PQ training, where each subvector only has 8 floats.

L2:

```
L2(simd,f32x8)          time:   [1.9428 ms 1.9509 ms 1.9561 ms]
                        change: [-26.987% -25.863% -24.769%] (p = 0.00 < 0.10)
                        Performance has improved.
```

Cosine:

```
Cosine(simd,8-f32) rng seed
                        time:   [2.8588 ms 2.8693 ms 2.8960 ms]
                        change: [-62.852% -62.407% -61.952%] (p = 0.00 < 0.10)
                        Performance has improved.
```